[feat] Create and destroy connection pool when app starts and shuts down.
