+++ .dockerignore
@@ -6,4 +6,4 @@ venv
 Dockerfile
 my_readme.txt
 README.md
-
+logs
