+++ Dockerfile
@@ -19,18 +19,16 @@ RUN pip install -r requirements.txt
 RUN mkdir -p /root/.ssh
 RUN ssh-keyscan github.com >> /root/.ssh/known_hosts
 
-COPY id_rsa /root/.ssh/id_rsa
+COPY secret/id_rsa /root/.ssh/id_rsa
 RUN chmod 600 /root/.ssh/id_rsa
 
 ARG BARCHART_API_TAG
 RUN if [ -z "$BARCHART_API_TAG" ]; then echo "BARCHART_API_TAG is required"; exit 1; fi
 RUN pip install git+ssh://git@github.com/hanchiang/barchart_api.git@$BARCHART_API_TAG && rm /root/.ssh/id_rsa
 
-RUN export PYTHONPATH=$(pwd)
-
-# copy project
 COPY . .
-RUN rm id_rsa
+RUN rm -rf "$(pwd)/secret"
+ENV PYTHONPATH "${PYTHONPATH}:$(pwd)"
 
 FROM base as dev
 CMD ["uvicorn", "--reload", "--app-dir", "src/server", "main:app"]
