+++ src/server/main.py
@@ -4,33 +4,28 @@ from typing import Optional
 
 from barchart_api import BarChartAPI
 
-import subprocess
-import os
-import asyncio
+from src.db.index import run_migration, start_postgres_connection_pool, stop_postgres_connection_pool
 
 app = FastAPI()
 
 options_api = BarChartAPI().options
 
-async def run_migration():
-    print('Running postgres migration scripts. Wait 5 seconds for timescale db to be ready')
-    await asyncio.sleep(5)
-    result = subprocess.run('piccolo migrations forwards market_data_piccolo --trace', cwd=os.getcwd(), timeout=30,
-                            shell=True, capture_output=True)
-    print(result)
-    print(f'stdout: {result.stdout}')
-    if result.stderr:
-        print(f'stderr: {result.stderr}')
-
 @app.on_event("startup")
 async def startup_event():
     print('FastAPI startup event')
+    await start_postgres_connection_pool()
     await run_migration()
+
+@app.on_event('shutdown')
+async def shutdown_event():
+    print('FastAPI shutdown event')
+    await stop_postgres_connection_pool()
 
 @app.get("/healthz")
 async def health_check():
     return {'data': 'market data is running!'}
 
+# TODO: Use an arg to determine whether to connect to API source or DB
 @app.get("/options/{symbol}")
 async def options_for_ticker(
     symbol: str, order_dir = '', expiration_type = '', expiration_date: Optional[date] = None,
@@ -44,6 +39,6 @@ async def options_for_ticker(
 
 @app.get("/options/{symbol}/expirations")
 async def options_expirations_for_ticker(symbol: str):
+    # TODO: cache
     data = await options_api.get_options_expirations_for_ticker(symbol=symbol)
-
     return {'data': data}
